requirements.txt
keras
matplotlib
tensorflow
numpy
colorlog
easydict
pyyaml
